pyproject.toml
[project]
name = "bonnici_portfolio"
version = "0.5.0"
description = "Bonnici's Portfolio"
authors = [
    {name = "Thomas Bonnici", email = "thomas.bonnici@icloud.com"},
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "toml>=0.10.2",
    "SQLAlchemy>=2.0.40",
    "jinja2>=3.1.6",
    "requests>=2.32.3",
    "python-multipart>=0.0.9",
    "user-agents>=2.2.0",
    "Pillow>=11.2.1",
    "colorthief>=0.2.1",
    "config-manager>=1.1.2"
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
front = {cmd = "uvicorn front.main:app --reload --host 0.0.0.0 --port 84"}
admin = {cmd = "uvicorn admin.main:app --reload --host 0.0.0.0 --port 85"}
generate_admin_password = {cmd = "python -m admin.cli generate_password"}
